[FIX] Edited files for more clarity
